move string parsing to its own mod
