+++ src/query/mod.rs
@@ -1,5 +1,6 @@
 pub mod query;
 pub mod selectors;
 
+mod strings;
 mod traversal;
 mod traversal_composites;
